// app/chatrpi/types.ts
/**
 * Copyright (c) 2020 - 2023 Persanix LLC. All rights reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export type LoadingStep =
    | "Begin"
    | "Loading samples"
    | "Training recognizer"
    | "Loading voice"
    | "Done"

export type LoadingError =
    | "No name samples found."
    | "No background samples found."
    | "Selected voice not found."

export type ConversationParty = "user" | "chatrpi";

export interface ConversationContent {
    party: ConversationParty;
    text: "gibberish" | string;
    time: number;
}

export type Conversation = ConversationContent[];

export type ConversationState =
    | "waiting"
    | "listening"
    | "thinking"
    | "responding"

export interface RingState {
    color: ColorHex,
    speed: number,
    position: Vector
}

export type Vector = [
    x: number,
    y: number,
    z: number
]

export type ColorHex = `#${ string }`;